==> 1.2 Download Posters.ipynb <==
#Import the required modules
import pandas as pd
import numpy as np
import io
import sys
import os.path
import urllib.request
from tqdm import tqdm
from os import listdir
from PIL import Image
import glob
import warnings
warnings.filterwarnings('ignore')

#Read movie data csv with poster url
df = pd.read_csv("Data/new.csv")
df.head()

#Shape of dataframe
df.shape

#Drop duplicates
df.drop_duplicates(subset="Movie_ID", inplace=True)

#Check null values(any column)
check_nan_in_df = df.isnull().values.any()
print (check_nan_in_df)

#Remove NaN valued entries
df= df.dropna(how='any')
df.shape

#df = df.sample(100)
#df['Movie_ID'] = df['Movie_ID'].astype(str)
#df = df.sort_values(by = 'Movie_ID')

#Method to download images using urllib
found = []
not_found = []
for index, row in tqdm(df.iterrows()):
    url = row['Poster_URL']
    imdb_id = row['Movie_ID']
    
    file_path = "Posters/" + str(imdb_id) + ".jpg"
    
    try:
        response = urllib.request.urlopen(url)
        data = response.read()
        file = open(file_path, 'wb')
        file.write(bytearray(data))
        file.close()
        found.append(file_path)
    except:
        not_found.append(imdb_id)
        
print("Number of posters not found = ", len(not_found))
print("Following IMDB_ID posters were not found::", not_found)

# Remove from dataframe those whose posters were not found
df = df[~df['Movie_ID'].isin(not_found)]
df.shape

# Check which downloaded images are corrupt and remove them 
bad_images = []
for file in glob.glob("Posters/*.jpg"):

    try:
        img = Image.open(file) # open image file
        img.verify() # verify its an image
    except (IOError, SyntaxError) as e:
        print('Bad file:', file) 

        bad_images.append(file)
        
print("Number of corrupt files:", len(bad_images))

# #Checking if all the imdb_id listed here actually have its poster image
# for file in glob.glob("Posters/*.jpg"):
#     image_list.append(file)
# print("Number of files found:", len(image_list))
df["Path"] = found

# Display and Save the updated Dataframe
df.to_csv('Data/data_with_path.csv', index=False)
display(df)
