==> 3.1 Preprocessing & Cleaning.ipynb <==
import pandas as pd
import numpy as np
import itertools
import warnings
from sklearn.preprocessing import LabelEncoder
from sklearn.model_selection import train_test_split, GridSearchCV
warnings.filterwarnings('ignore')
import re
import nltk
import string
from nltk.corpus import stopwords
from nltk import word_tokenize, FreqDist
from nltk.stem import WordNetLemmatizer
import spacy

df = pd.read_csv("Data/new.csv")
df = df[["Title", "Genre", "Synopsis"]]
df = df.assign(Genre=df.Genre.str.split(",")).explode('Genre')
df.to_csv('Data/test.csv', index = False)
df.head()

#defining the function to remove punctuation
def remove_punctuation(text):
    punctuationfree="".join([i for i in text if i not in string.punctuation])
    return punctuationfree

#defining function for tokenization
def tokenization(text):
    tokens = re.split('W+',text)
    return tokens

#Stop words present in the library
stopwords = nltk.corpus.stopwords.words('english')
stopwords[0:10]
['i', 'me', 'my', 'myself', 'we', 'our', 'ours', 'ourselves', 'you', "you're"]

#defining the function to remove stopwords from tokenized text
def remove_stopwords(text):
    output= [i for i in text if i not in stopwords]
    return output


#defining the object for Lemmatization
wordnet_lemmatizer = WordNetLemmatizer()

#defining the function for lemmatization
def lemmatizer(text):
    lemm_text = [wordnet_lemmatizer.lemmatize(word) for word in text]
    return lemm_text

df['Synopsis']= df['Synopsis'].apply(lambda x:remove_punctuation(x))
df['Synopsis']= df['Synopsis'].apply(lambda x: x.lower())
df['Synopsis']= df['Synopsis'].apply(lambda x: tokenization(x))
df['Synopsis']= df['Synopsis'].apply(lambda x:remove_stopwords(x))
df['Synopsis']= df['Synopsis'].apply(lambda x:lemmatizer(x))
df['Synopsis']= df['Synopsis'].apply(str)
df.head()

df.to_csv('Data/cleaned.csv', index = False)
