==> 4.2 Model Training.ipynb <==
#Importing all Required Libraries
import numpy as np
import pandas as pd
from sklearn.model_selection import train_test_split
from sklearn import metrics
from sklearn.linear_model import LogisticRegression
from sklearn.naive_bayes import MultinomialNB
from sklearn.pipeline import Pipeline
from sklearn.tree import DecisionTreeClassifier
from sklearn.ensemble import RandomForestClassifier
from sklearn.feature_extraction.text import CountVectorizer
import pickle
import warnings
warnings.filterwarnings('ignore')

df = pd.read_csv("Data/rev.csv")
df.head()

X = df["Review"].values
y = df["Rating"].values

#Spliting of Dataset into tarining and testing datasets
review_train, review_test, label_train, label_test = train_test_split(X, y, test_size=0.3, random_state=101)
new = df[['Review','Rating']].copy()

#Applying bag of words Text Feature Extraction with Pipeline using countvectorizer, and classifer - LogisticRegression
pipeline1 = Pipeline([
    ('bag_of_words', CountVectorizer()),
    ('classifier', LogisticRegression(solver='newton-cg', multi_class='multinomial'))
])

pipeline1.fit(review_train, label_train)
pip_pred1 = pipeline1.predict(review_test)
print(metrics.classification_report(label_test,pip_pred1))

filename = 'model_lr.pk'
with open('Models/'+filename, 'wb') as file:
    pickle.dump(pipeline1, file) 

#Applying bag of words Text Feature Extraction with Pipeline using countvectorizer, and classifer - MultinomialNB
pipeline2 = Pipeline([
    ('bag_of_words', CountVectorizer()),
    ('classifier', MultinomialNB())
])
pipeline2.fit(review_train, label_train)
pip_pred2 = pipeline2.predict(review_test)
print(metrics.classification_report(label_test, pip_pred2))

filename = 'model_mnb.pk'
with open('Models/'+filename, 'wb') as file:
    pickle.dump(pipeline2, file) 

#Applying bag of words Text Feature Extraction with Pipeline using countvectorizer, and classifer - DecisionTreeClassifier
pipeline3 = Pipeline([
    ('bag_of_words', CountVectorizer()),
    ('classifier', DecisionTreeClassifier())
])

pipeline3.fit(review_train, label_train)
pip_pred3 = pipeline3.predict(review_test)
print(metrics.classification_report(label_test, pip_pred3))

filename = 'model_dt.pk'
with open('Models/'+filename, 'wb') as file:
    pickle.dump(pipeline3, file) 

#Applying bag of words Text Feature Extraction with Pipeline using countvectorizer, and classifer - RandomForestClassifier
pipeline4 = Pipeline([
    ('bag_of_words', CountVectorizer()),
    ('classifier', RandomForestClassifier())
])

pipeline4.fit(review_train, label_train)
pip_pred4 = pipeline4.predict(review_test)
print(metrics.classification_report(label_test, pip_pred4))

filename = 'model_rf.pk'
with open('Models/'+filename, 'wb') as file:
    pickle.dump(pipeline4, file)
