==> 4.3 Classification & Validation.ipynb <==
#Importing all Required Libraries
import numpy as np
import pandas as pd
import random
from sklearn.metrics import accuracy_score
import glob
import pickle
import warnings
warnings.filterwarnings('ignore')
import os

df = pd.read_csv("Data/new.csv")
df.head()

folder = os.fsencode("Reviews")
#Declaring a list which contains the filenames of all the files in that directory
filenames = []
#Traversing the folder, and appending all the files that ends with csv, to the list
for file in os.listdir(folder):
    filename = os.fsdecode(file)
    if filename.endswith( ('.csv') ):
        filenames.append(filename)
        
display(filenames)

sliced_strings = []
for string in filenames:
    sliced_strings.append(string[:-4])
sliced_strings

df.sort_values(by='Title', ascending=True, inplace=True)
df = df[df['Title'].isin(sliced_strings)]
df.head()

models = ["model_lr.pk", "model_dt.pk", "model_rf.pk", "model_mnb.pk"]
pipeline = []

for i in models:
    with open("Models/" + i ,'rb') as f:
        pipeline.append(pickle.load(f))

d = random.randint(0, len(sliced_strings))
p = pd.read_csv("Reviews//"+sliced_strings[d]+".csv")
display(p.head())
p = p[['Rating','Review']].copy()
p['Rating'] = p['Rating'].astype('string')
p = p.dropna()
p['Rating'] = pd.to_numeric(p['Rating'])
    # p["Rating"]=p["Rating"]/2
p['Rating']=p['Rating'].astype(int)
 
X = p['Review'].copy()
y = p['Rating'].copy()

pre = []
for i in range(4):
    pred = pipeline[i].predict(X)
    print("Model: ", models[i])
    print(accuracy_score(y, pred))
    avg = sum(pred)/len(pred)
    pre.append(avg)
mean = sum(pre)/len(pre)
print("\nMovie: ", sliced_strings[d])
print("Predicted Rating: ", mean)

res = df[df['Title'] == sliced_strings[d]]
print("Orignal Rating: ", res.iat[0, 10])

X
